requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
paho-mqtt==1.6.1
pyls-spyder==0.4.0
PyMySQL==1.1.0
SQLAlchemy==2.0.23
watchdog==3.0.0
webencodings==0.5.1
Werkzeug==3.0.1
psutil
cryptography
